+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.10.0
 
+- Win: Suppress asyncio transport \_\_del\_\_ exceptions (ValueError from closed pipes) during interpreter shutdown
 - Protect against Ctrl-C/KeyboardInterrupt in critical steps
 - Don't print microseonds in the log
 - Ctrl-C/KeyboardInterrupt results in exit code 130